[PDI-17629] "Wait For" Job Entry step doesn't have a consistent behaviour for the set "Maximum timeout" value

+++ engine/src/main/java/org/pentaho/di/job/entries/delay/JobEntryDelay.java
@@ -2,7 +2,7 @@
  *
  * Pentaho Data Integration
  *
- * Copyright (C) 2002-2017 by Hitachi Vantara : http://www.pentaho.com
+ * Copyright (C) 2002-2018 by Hitachi Vantara : http://www.pentaho.com
  *
  *******************************************************************************
  *
@@ -159,9 +159,9 @@ public class JobEntryDelay extends JobEntryBase implements Cloneable, JobEntryIn
 
     try {
       // starttime (in seconds ,Minutes or Hours)
-      long timeStart = System.currentTimeMillis() / Multiple;
+      double timeStart = (double) System.currentTimeMillis() / (double) Multiple;
 
-      long iMaximumTimeout = Const.toInt( getRealMaximumTimeout(), Const.toInt( DEFAULT_MAXIMUM_TIMEOUT, 0 ) );
+      double iMaximumTimeout = Const.toInt( getRealMaximumTimeout(), Const.toInt( DEFAULT_MAXIMUM_TIMEOUT, 0 ) );
 
       if ( isDetailed() ) {
         logDetailed( BaseMessages.getString( PKG, "JobEntryDelay.LetsWaitFor.Label", iMaximumTimeout, Waitscale ) );
@@ -181,7 +181,7 @@ public class JobEntryDelay extends JobEntryBase implements Cloneable, JobEntryIn
       //
       while ( continueLoop && !parentJob.isStopped() ) {
         // Update Time value
-        long now = System.currentTimeMillis() / Multiple;
+        double now = (double) System.currentTimeMillis() / (double) Multiple;
 
         // Let's check the limit time
         if ( ( iMaximumTimeout >= 0 ) && ( now >= ( timeStart + iMaximumTimeout ) ) ) {
